:sparkles: Add a RemoteWorker class to handle remote operations.
This moves a lot of the code that was in index.js into RemoteWorker.js.
This file contains an encapsulation of the information needed to establish
an SSH connection to a remote server, allowing us to better test this code,
as it's more isolated from the main command line interface.